129: split services/sessions.rs by concern (PR-G4)
Splits the 2693-line services/sessions.rs (1290 non-test) into three
files under services/sessions/, each well under rust.md § 13's ~500-line
non-test threshold:

- sessions/lifecycle.rs (~450 non-test): create / join / leave /
  host-transfer / close-stale / list-active / get-active /
  REJOIN_GRACE_MINUTES / participant-in-active-session check.
- sessions/detail.rs (~370 non-test): get_session_detail / list_races
  with their supporting types (SessionDetail, ParticipantInfo,
  SessionRaceInfo, RaceSubmission, RaceInfo) and per-query helpers
  (load_host_username, load_participants,
  load_current_race_with_submissions, load_race_history).
- sessions/races.rs (~390 non-test): next_track / skip_turn /
  get_pending_races / skip_pending_race.

The new services/sessions.rs is a thin module root that declares the
three submodules and re-exports their public items, so callers continue
to reach everything via crate::services::sessions::<name>.

Splits along the concern boundaries identified in Cowork's
docs/designs/2026-04-15-rust-audit.md. Tests follow their production
code into each submodule's #[cfg(test)] mod tests.

git mv was used to move the original sessions.rs to sessions/races.rs
(the largest extract at ~48% content similarity). Because the split is
1→3, default `git log --follow` (50% similarity threshold) will not
trace across detail.rs and lifecycle.rs. Use
`git log --follow -M30 -C` or `git blame -C -C` for function-level
history across the split.

Per docs/compliance-plan.md PR-G4 / Issue #129. No API or schema change.
All 56 affected unit tests (now distributed across the three submodules)
plus the existing 19 session integration tests pass unchanged.
